Implement selector for Haiku
Since Haiku supports neither `epoll` nor `kqueue`, uses `poll`.
Unfortunately the `poll` API is not stateful like the others, so
Selector has to keep track of some additional state (what fds are
registered for which events).

As a consequence, changes were required to a bunch of other files to
make references mutable when targetting Haiku. Hopefully there's a
better way to implement this, because it adds a lot of code.

HACK: With poll, there is no associated file descriptor, so `as_raw_fd`
will always return 1. No idea yet if this causes problems.

TODO: `derive(Debug)` doesn't work for `Selector` since it has a
`Vec<libc::pollfd>` member and the `libc` crate doesn't `impl Debug` for
`pollfd`. Implement a proper `Debug` trait for `Selector` that formats
the `fds` member.